splitting up analysis into submodules.
